+++ lib/transpareo/md2pdf/highlighter.rb
@@ -18,7 +18,7 @@ module Transpareo
         'console' => 'shell',
         'docker' => 'docker',
         'tsx' => 'typescript',
-        'jsx' => 'javascript'
+        'jsx' => 'javascript',
       }.freeze
 
       module_function